public

+++ models/OpenAIGPT.py
@@ -29,10 +29,10 @@ class OpenAIGPT(BaseModel):
 
 
 class OpenAIGPTWithChatTemplate(BaseModel):
-    def __init__(self, model_name: str = "openai-community/openai-gpt", *args, **kwargs):
+    def __init__(self, model_name: str = "openai-community/openai-gpt", dtype=torch.float16, *args, **kwargs):
         # assert model_name in openai_all_gpts
         tokenizer = AutoTokenizer.from_pretrained(model_name)
-        model = AutoModelForCausalLM.from_pretrained(model_name)
+        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=dtype)
         conv = get_conv_template("vicuna_v1.1")  # personally I like this template most.
         # I also believe an only pretrained model can be able to do conversation using this template
         super(OpenAIGPTWithChatTemplate, self).__init__(model, tokenizer, conv, *args, **kwargs)
